some cool changes

+++ app/src/main/java/com/igorvorobiov/sunshine/ForecastItem.java
@@ -28,11 +28,11 @@ public class ForecastItem {
         return description;
     }
 
-    public Double getMax(){
-        return max;
+    public Integer getMax(){
+        return max.intValue();
     }
 
-    public Double getMin(){
-        return min;
+    public Integer getMin(){
+        return min.intValue();
     }
 }
